auth/requirements.txt
aiohttp>=3.7.2
fastapi>=0.61.2
pycryptodome >= 3.9.9
pydantic>=1.7.2
python-multipart>=0.0.5
typing>=3.7.4.3
ujson>=4.0.1
uvicorn>=0.12.2
uvloop>=0.14.0; platform_system != "Windows"
